Move classes from the main script to library
